+++ changelog.md
@@ -1,4 +1,16 @@
 # Change Log
 
+## Version 0.2.0
+New Features:
+- Added chat messages when activities are completed
+- Sped up the flow for creating new downtime activities
+- Added settings to configure default options for both Simple and Ability Check progression systems
+
+Fixes:
+- Updated the text for the old completion threshold setting to accurate describe its new function
+
+Behind The Scenes:
+- Refactored some things to allow for easier additions of more progression systems
+
 ## Version 0.1.0
 Initial Release